requirements.txt
pytorch==1.9.0
pandas==1.1.3
numpy==1.21.2
matplotlib==3.3.3
scikit-learn==0.23.2